FlaskWebProject/requirements.txt
asgiref==3.3.4
backoff==1.10.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
googleapis-common-protos==1.52.0
grpcio==1.57.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
opencensus-proto==0.1.0
opentelemetry-api==1.10a0
opentelemetry-exporter-opencensus==0.20b0
opentelemetry-exporter-otlp-proto-grpc==1.10a0
opentelemetry-instrumentation==0.20b0
opentelemetry-instrumentation-flask==0.20b0
opentelemetry-instrumentation-logging==0.20b0
opentelemetry-instrumentation-requests==0.20b0
opentelemetry-instrumentation-wsgi==0.20b0
opentelemetry-proto==1.0.0
opentelemetry-sdk==1.10a0
opentelemetry-semantic-conventions==0.20b0
opentelemetry-util-http==0.20b0
protobuf==3.15.8
psutil==5.8.0
requests==2.25.1
six==1.15.0
urllib3==1.26.4
Werkzeug==1.0.1
wrapt==1.12.1
